node now too

+++ scripts/bootstrap_node.sh
@@ -1,5 +1,17 @@
 #! /bin/bash
 
+# Check if both arguments are provided
+if [ $# -ne 2 ]; then
+    echo -e "\x1b[31mError: Please provide both address and port arguments.\x1b[0m"
+    echo -e "\x1b[33mUsage: $0 <address> <port>\x1b[0m"
+    echo -e "\x1b[33mExample: $0 127.0.0.1 8080\x1b[0m"
+    exit 1
+fi
+
+# Store command line arguments
+ADDRESS=$1
+PORT=$2
+
 echo -e "\x1b[33mChecking for existing tinyweather-node service file...\x1b[0m"
 if [[ -f /etc/systemd/system/tinyweather-node.service ]]; then 
     systemctl stop tinyweather-node.service
@@ -10,63 +22,47 @@ if [[ -f /etc/systemd/system/tinyweather-node.service ]]; then
 else
     echo -e "\x1b[32mNo existing tinyweather-node service file found. Skipping removal steps.\x1b[0m"
 fi
-
 echo -e "\x1b[32mBeginning the bootstrapping of tinyweather-node\x1b[0m"
 echo -e "\x1b[33mBuilding tinyweather-node with 'zig build'...\x1b[0m"
 zig build 
-
-
 echo -e "\x1b[33mPreparing /opt/tinyweather directory...\x1b[0m"
 mkdir -p /opt/tinyweather
-
 echo -e "\x1b[33mRemoving existing tinyweather-node binary from /opt/tinyweather (if it exists)...\x1b[0m"
 rm -f /opt/tinyweather/tinyweather-node
-
 echo -e "\x1b[33mMoving newly built tinyweather-node executable to /opt/tinyweather\x1b[0m"
 mv ./zig-out/bin/tinyweather-node /opt/tinyweather
-
 if [[ -d .venv ]]; then 
   rm -rf .venv
 fi
-
 python3 -m venv .venv
 source .venv/bin/activate
 pip install adafruit-circuitpython-bme680 adafruit-blinka RPi.GPIO
-
 if [[ -d /opt/tinyweather/.venv/ ]]; then 
   rm -rf /opt/tinyweather/.venv
 fi
-
 mv ./.venv /opt/tinyweather
-
 echo -e "\x1b[33mCreating systemd service file at /etc/systemd/system/tinyweather-node.service...\x1b[0m"
 touch /etc/systemd/system/tinyweather-node.service
-
 echo -e "\x1b[33mWriting service configuration to /etc/systemd/system/tinyweather-node.service...\x1b[0m"
 echo "
 [Unit]
 Description=Tinyweather Node Service
 After=network.target
-
 [Service]
 Type=simple
 Restart=always
 RestartSec=5s
-ExecStart=/opt/tinyweather/tinyweather-node --address 127.0.0.1 --port 8080 
+ExecStart=/opt/tinyweather/tinyweather-node --address $ADDRESS --port $PORT 
 WorkingDirectory=/opt/tinyweather
-Environment="PATH=/opt/tinyweather/.venv/bin:/usr/bin"
-
+Environment=\"PATH=/opt/tinyweather/.venv/bin:/usr/bin\"
 [Install]
 WantedBy=multi-user.target" >> /etc/systemd/system/tinyweather-node.service
 echo -e "\x1b[32mService configuration written to /etc/systemd/system/tinyweather-node.service\x1b[0m"
-
 echo -e "\x1b[33mReloading systemd daemon to recognize the new tinyweather-node service...\x1b[0m"
 systemctl daemon-reload
 echo -e "\x1b[32mSystemd daemon reloaded successfully.\x1b[0m"
-
 echo -e "\x1b[33mStarting the tinyweather-node service...\x1b[0m"
 systemctl start tinyweather-node
 systemctl enable tinyweather-node
 echo -e "\x1b[32mtinyweather-node service started.\x1b[0m"
-
 systemctl status tinyweather-node
